backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pytest==8.2.0
httpx==0.27.0
openai==1.30.1
python-dotenv==1.0.1
